// types.ts
export enum Category {
  WORK = 'Work',
  PERSONAL = 'Personal',
  URGENT = 'Urgent',
  LEARNING = 'Learning',
  HEALTH = 'Health'
}

export interface Todo {
  id: string;
  text: string;
  completed: boolean;
  category: Category;
  createdAt: number;
  subtasks?: SubTask[];
  isAI?: boolean; // If generated by AI
}

export interface SubTask {
  id: string;
  text: string;
  completed: boolean;
}

export type FilterType = 'ALL' | 'ACTIVE' | 'COMPLETED';